packages/paper_elements/src/web-animations-next/test/blink/get-animation-players.html

<!DOCTYPE html><script src="testharness/testharness.js"></script><script src="testharness/testharnessreport.js"></script><div id=container><div id=element></div></div><script>var container=document.getElementById("container");var element=document.getElementById("element");test(function(){assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0);var b=element.animate([],1000);assert_equals(document.timeline.getAnimationPlayers().length,1);assert_equals(document.timeline.getAnimationPlayers()[0],b);var a=container.animate([],1000);assert_equals(document.timeline.getAnimationPlayers().length,2);assert_equals(document.timeline.getAnimationPlayers()[0],b);assert_equals(document.timeline.getAnimationPlayers()[1],a);b.finish();assert_equals(document.timeline.getAnimationPlayers().length,1);assert_equals(document.timeline.getAnimationPlayers()[0],a);a.finish();assert_equals(document.timeline.getAnimationPlayers().length,0)},"Timeline getAnimationPlayers()");test(function(){assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0);var b=element.animate([],1000);assert_equals(document.timeline.getAnimationPlayers().length,1);assert_equals(document.timeline.getAnimationPlayers()[0],b);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,1);assert_equals(element.getAnimationPlayers()[0],b);var a=container.animate([],1000);assert_equals(document.timeline.getAnimationPlayers().length,2);assert_equals(document.timeline.getAnimationPlayers()[0],b);assert_equals(document.timeline.getAnimationPlayers()[1],a);assert_equals(container.getAnimationPlayers().length,1);assert_equals(container.getAnimationPlayers()[0],a);assert_equals(element.getAnimationPlayers().length,1);assert_equals(element.getAnimationPlayers()[0],b);b.finish();assert_equals(document.timeline.getAnimationPlayers().length,1);assert_equals(document.timeline.getAnimationPlayers()[0],a);assert_equals(container.getAnimationPlayers().length,1);assert_equals(container.getAnimationPlayers()[0],a);assert_equals(element.getAnimationPlayers().length,0);a.finish();assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0)},"Animatable getAnimationPlayers()");test(function(){assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0);var a=element.animate([],{duration:1000,delay:500});assert_equals(document.timeline.getAnimationPlayers().length,1);assert_equals(document.timeline.getAnimationPlayers()[0],a);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,1);assert_equals(element.getAnimationPlayers()[0],a);a.finish();assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0)},"getAnimationPlayers() with delays");</script>
